Refactor: DTO record 클래스에서 @Builder 사용 지양
- Builder 대신 생성자를 통해 객체 생성하도록 수정
