+++ src/main/java/com/sprint/mission/discodeit/dto/service/binaryContent/BinaryContentDTO.java
@@ -1,11 +1,8 @@
 package com.sprint.mission.discodeit.dto.service.binaryContent;
 
-import lombok.Builder;
-
 import java.time.Instant;
 import java.util.UUID;
 
-@Builder
 public record BinaryContentDTO(
         UUID id,
         Instant createdAt,
