+++ src/main/java/com/sprint/mission/discodeit/dto/service/binaryContent/CreateBinaryContentParam.java
@@ -1,8 +1,5 @@
 package com.sprint.mission.discodeit.dto.service.binaryContent;
 
-import lombok.Builder;
-
-@Builder
 public record CreateBinaryContentParam(
         String filename,
         String path,
