+++ src/main/java/com/sprint/mission/discodeit/dto/service/channel/ChannelDTO.java
@@ -1,12 +1,11 @@
 package com.sprint.mission.discodeit.dto.service.channel;
 
 import com.sprint.mission.discodeit.entity.ChannelType;
-import lombok.Builder;
 
 import java.time.Instant;
 import java.util.UUID;
 
-@Builder
+
 public record ChannelDTO(
         UUID id,
         Instant createdAt,
