+++ src/main/java/com/sprint/mission/discodeit/dto/service/message/CreateMessageParam.java
@@ -1,7 +1,5 @@
 package com.sprint.mission.discodeit.dto.service.message;
 
-import lombok.Builder;
-
 import java.util.List;
 import java.util.UUID;
