+++ src/main/java/com/sprint/mission/discodeit/dto/service/message/MessageDTO.java
@@ -1,12 +1,10 @@
 package com.sprint.mission.discodeit.dto.service.message;
 
-import lombok.Builder;
-
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
 
-@Builder
+
 public record MessageDTO(
         UUID id,
         Instant createdAt,
