+++ src/main/java/com/sprint/mission/discodeit/dto/service/readStatus/CreateReadStatusParam.java
@@ -1,10 +1,7 @@
 package com.sprint.mission.discodeit.dto.service.readStatus;
 
-import lombok.Builder;
-
 import java.util.UUID;
 
-@Builder
 public record CreateReadStatusParam(
         UUID userId,
         UUID channelId
