+++ src/main/java/com/sprint/mission/discodeit/dto/service/user/UserDTO.java
@@ -1,11 +1,9 @@
 package com.sprint.mission.discodeit.dto.service.user;
 
-import lombok.Builder;
 import java.time.Instant;
 import java.util.UUID;
 
 
-@Builder
 public record UserDTO(
         UUID id,
         UUID profileId,
